First pass at updates to pyLHF interface
